+++ CHANGELOG.md
@@ -71,6 +71,7 @@
 - Updated the `make.sh` script to automatically infer the version to use from the source's `package.json`.
 - Simplified the process of maintaining the revealJS templates, added a few other goodies. The command `reveal:build` will now re-build the full revealJS templates with the installed revealJS version.
 - The Citeproc-provider now logs all errors that prevent a successful boot to the console.
+- Replaced the sidebar with a Vue.js component.
 
 # 1.3.0